Modify AFT tow program
Reduces jerk, acceleration, and velocity for carriage on tow-back to prevent AFT shaft from being loaded in tension and moving out of typical position in nacelle. Attempt to change PTP direction by assigning a negative velocity beforehand so turbine always moves in CCW direction. This is to account for lash in the shaft coupling that might change '0' if moved in the opposite direction.