Converge on AFT motion program and rename temp. channel (#96)
* AFT axis to 0 after carriage is back at right limit

* Wait 2 s after HALT(6) command

* PTP(6) before tow back

* Modify AFT tow program

Reduces jerk, acceleration, and velocity for carriage on tow-back to prevent AFT shaft from being loaded in tension and moving out of typical position in nacelle. Attempt to change PTP direction by assigning a negative velocity beforehand so turbine always moves in CCW direction. This is to account for lash in the shaft coupling that might change '0' if moved in the opposite direction.

* Increase tow-back velocity

* yaskawa_temp --> water_temp

Using extra RTD for water temperature instead of monitoring temperature inside Yaskawa cabinet

* Update AFT NI plots for water temp. instead

* Increase DEC(6)

* Increase DEC(6)

* Reduce JERK(5) by 25%

* Decrease DEC(5) to 0.3

* Decrease DEC(5) to 0.15

* DEC(5) back to 0.5, reduce JERK on both axes

* Decrease DEC(6) to 1/10th ACC(6)

* Decrease JERK(6)

* Decrease JERK(5)

* Decrease both JERK params.

* Return back to normal JERK param.

* Finalize AFT motion program

* Update turbinedaq/acsprgs.py

---------

Co-authored-by: Pete Bachant <petebachant@gmail.com>